SectionMapModel - Finished adding Section mapping.
SectionMapModel - Replaced original model with simple 1-to-1 mappings
between local and remote Sections. Users enter the index of the target.
SerialDeviceController - Auto load (or initialize) Section map on
initialization.
DeviceControlWidget - Fixed Section map check during `run_cue()`.
DeviceControlWidget - Fixed auto loading and auto saving Section map
with device.

+++ src/controller/serialdevicethreadcontroller.cpp
@@ -3,7 +3,7 @@
 #include "serialdevicethreadcontroller.h"
 
 namespace PixelMaestroStudio {
-	SerialDeviceThreadController::SerialDeviceThreadController(SerialDeviceController *serial_device, const char *out, uint16_t size) : QThread(nullptr) {
+	SerialDeviceThreadController::SerialDeviceThreadController(SerialDeviceController *serial_device, const char *out, int size) : QThread(nullptr) {
 		this->output_.append(out, size);
 		this->serial_device_ = serial_device;
 	}
